requirements.txt
pyTelegramBotAPI
python-dotenv
psycopg2-binary
sqlalchemy